Make extraction and host projection explicit
