--- src/TinyDispatcher.SourceGen/Generator/Composition/ContextComposition.cs
@@ -1,10 +1,0 @@
-using TinyDispatcher.SourceGen.Generator.Models;
-
-namespace TinyDispatcher.SourceGen.Generator.Composition;
-
-internal readonly record struct ContextComposition(
-    HostContextInfo HostContext,
-    string ContextTypeFqn,
-    PipelineConfig ThisAssemblyPipeline,
-    DiscoveryResult HostDiscovery,
-    PipelineConfig HostPipeline);
