+++ src/TinyDispatcher.SourceGen/Generator/Composition/ValidationInputComposer.cs
@@ -6,26 +6,24 @@ namespace TinyDispatcher.SourceGen.Generator.Composition;
 internal sealed class ValidationInputComposer
 {
     public ImmutableArray<ContextValidationInput> Compose(
-        ImmutableArray<ContextComposition> contexts,
-        ImmutableArray<ContextGenerationInput> generationContexts)
+        ImmutableArray<HostContextProjection> contexts)
     {
         var validationInputs = ImmutableArray.CreateBuilder<ContextValidationInput>(contexts.Length);
 
         for (var i = 0; i < contexts.Length; i++)
         {
-            validationInputs.Add(BuildValidationContext(contexts[i], generationContexts[i]));
+            validationInputs.Add(BuildValidationContext(contexts[i]));
         }
 
         return validationInputs.ToImmutable();
     }
 
     private static ContextValidationInput BuildValidationContext(
-        ContextComposition context,
-        ContextGenerationInput generationContext)
+        HostContextProjection context)
     {
         return new ContextValidationInput(
             BootstrapCalls: context.HostContext.UseTinyDispatcherCalls,
             ThisAssemblyPipeline: context.ThisAssemblyPipeline,
-            GenerationInput: generationContext);
+            GenerationInput: context.GenerationInput);
     }
 }
