+++ src/TinyDispatcher.SourceGen/Generator/Models/GeneratorExtraction.cs
@@ -1,8 +1,5 @@
-using System.Collections.Immutable;
-
 namespace TinyDispatcher.SourceGen.Generator.Models;
 
 internal sealed record GeneratorExtraction(
-    DiscoveryResult Discovery,
-    ReferencedAssemblyContributions ReferencedContributions,
-    ImmutableArray<ContextPipelineConfig> ContextPipelines = default);
+    ThisAssemblyExtraction ThisAssembly,
+    ReferencedAssemblyContributions ReferencedContributions);
