+++ src/TinyDispatcher.SourceGen/Generator/Models/HostContextProjection.cs
@@ -1,0 +1,6 @@
+namespace TinyDispatcher.SourceGen.Generator.Models;
+
+internal sealed record HostContextProjection(
+    HostContextInfo HostContext,
+    PipelineConfig ThisAssemblyPipeline,
+    ContextGenerationInput GenerationInput);
