+++ src/TinyDispatcher.SourceGen/Generator/Models/HostGenerationComposition.cs
@@ -5,4 +5,4 @@ namespace TinyDispatcher.SourceGen.Generator.Models;
 internal sealed record HostGenerationComposition(
     DiscoveryResult Discovery,
     ReferencedAssemblyContributions ReferencedContributions,
-    ImmutableArray<ContextGenerationInput> Contexts);
+    ImmutableArray<HostContextProjection> Contexts);
