+++ src/TinyDispatcher.SourceGen/Generator/Models/ThisAssemblyExtraction.cs
@@ -1,0 +1,7 @@
+using System.Collections.Immutable;
+
+namespace TinyDispatcher.SourceGen.Generator.Models;
+
+internal sealed record ThisAssemblyExtraction(
+    DiscoveryResult Discovery,
+    ImmutableArray<ContextPipelineConfig> ContextPipelines = default);
